requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.5
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
openpyxl==3.1.5
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
ruff==0.12.9
sqlparse==0.5.3
uritemplate==4.2.0
